+++ core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Add `ConnectionLimit::with_max_established` (see [PR 2137]).
 
-[PR 2137]: https://github.com/libp2p/rust-libp2p/pull/2137/
+- Add `Keypair::to_protobuf_encoding` (see [PR 2142]).
 
 - Change `PublicKey::into_protobuf_encoding` to `PublicKey::to_protobuf_encoding` (see [PR 2145]).
 
@@ -13,6 +13,8 @@
 - Add `From<&PublicKey> for PeerId` (see [PR 2145]).
 
 [PR 2145]: https://github.com/libp2p/rust-libp2p/pull/2145
+[PR 2142]: https://github.com/libp2p/rust-libp2p/pull/2142
+[PR 2137]: https://github.com/libp2p/rust-libp2p/pull/2137/
 
 # 0.29.0 [2021-07-12]